# src/api/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "بات هوشمند اینستاگرام API",
    "description": "API برای مدیریت و کنترل بات هوشمند اینستاگرام با رفتار انسانی",
    "version": "1.0.0",
    "contact": {
      "name": "پشتیبانی",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "سرور توسعه محلی"
    }
  ],
...
  "paths": {
...
    "/api/stats/account": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Get account statistics",
        "description": "Retrieve general statistics about the Instagram account",
        "responses": {
          "200": {
            "description": "Account statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "daily": {
                          "type": "object",
                          "properties": {
                            "likes": {
                              "type": "integer",
                              "example": 45
                            },
                            "comments": {
                              "type": "integer",
                              "example": 12
                            },
                            "follows": {
                              "type": "integer",
                              "example": 8
                            },
                            "unfollows": {
                              "type": "integer",
                              "example": 5
                            },
                            "directMessages": {
                              "type": "integer",
                              "example": 10
                            },
                            "storyViews": {
                              "type": "integer",
                              "example": 30
                            },
                            "total": {
                              "type": "integer",
                              "example": 110
                            }
                          }
                        },
                        "weekly": {
                          "type": "object",
                          "properties": {
                            "interactions": {
                              "type": "integer",
                              "example": 550
                            }
                          }
                        },
                        "monthly": {
                          "type": "object",
                          "properties": {
                            "interactions": {
                              "type": "integer",
                              "example": 2200
                            }
                          }
                        },
                        "follows": {
                          "type": "object",
                          "properties": {
                            "active": {
                              "type": "integer",
                              "example": 150
                            },
                            "followBacks": {
                              "type": "integer",
                              "example": 84
                            },
                            "followBackRate": {
                              "type": "number",
                              "example": 56.0
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Active account not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "No active account found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}